requirements.txt
fastapi[standard]
uvicorn
asyncpg
python-dotenv
sqlalchemy
pyjwt
passlib[bcrypt]
psycopg2-binary
jinja2
pygments
tubectrl